updated functions

+++ portfolioML/makedir.py
@@ -1,4 +1,5 @@
 import argparse
+import sys
 import logging
 import os
 import shutil
@@ -15,7 +16,8 @@ def go_up(level_up):
 
     Returns
     --------
-    None
+    path : str
+        Path of the directory at level you chose from the current directory.
     '''
     if level_up == 0:
         path = os.getcwd()
@@ -53,32 +55,32 @@ def smart_makedir(name_dir, level_up=0):
 
     '''
 
+    separator = '/'
     if level_up == 0:
-        path = go_up(0) + "/" + name_dir
+        path = separator.join([go_up(0), name_dir])
     if level_up == 1:
-        path = go_up(1) + "/" + name_dir
+        path = separator.join([go_up(1), name_dir])
     if level_up == 2:
-        path = go_up(2) + "/" + name_dir
+        path = separator.join([go_up(2), name_dir])
     if level_up == 3:
-        path = go_up(3) + "/" + name_dir
+        path = separator.join([go_up(3), name_dir])
     if level_up == 4:
-        path = go_up(4) + "/" + name_dir
+        path = separator.join([go_up(4), name_dir])
 
     if os.path.exists(path):
-        logging.info(
-            f"Path '{path}' already exists, it will be overwritten \n")
-        # Remove all the files in case they already exist
-        shutil.rmtree(path)
+        answer = input('Path already exists. Do you want to overwrite the files? [y/n] ')
+        if answer == 'y':
+            # Remove all the files in case they already exist
+            shutil.rmtree(path)
+        else:
+            logging.info("I am quitting \n")
+            sys.exit()
     os.makedirs(path)
     logging.info(f"Successfully created the directory '{path}' \n")
 
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Easy way to create folders')
-    parser.add_argument('name_directory', type=str,
-                        help='Name of the directory that will be created')
-    parser.add_argument('-level_up', default=0,
-                        help='How many step up yo want from the current directory')
     parser.add_argument("-log", "--log", default="info",
                         help=("Provide logging level. Example --log debug', default='info"))
     args = parser.parse_args()
@@ -88,7 +90,4 @@ if __name__ == '__main__':
               'warning': logging.WARNING,
               'info': logging.INFO,
               'debug': logging.DEBUG}
-
     logging.basicConfig(level=levels[args.log])
-
-    smart_makedir(args.name_directory, args.level_up)
